Low-Level ARIMA function for translating modeltime to forecast

Low-Level ARIMA function for translating modeltime to forecast

garch_stan_fit_impl(
  x,
  y,
  s = 1,
  k = 1,
  h = 1,
  p = 0,
  q = 0,
  genT = FALSE,
  asym = "none",
  chains = 4,
  iter = 2000,
  warmup = iter/2,
  adapt.delta = 0.9,
  tree.depth = 10,
  seed = NULL,
  ...
)

Arguments

x
A dataframe of xreg (exogenous regressors)

y
A numeric vector of values to fit

s
garch_order

k
arch_order

h
mgarch_order

p
The order of the non-seasonal auto-regressive (AR) terms. Often denoted "p" in pdq-notation.

q
The order of the non-seasonal moving average (MA) terms. Often denoted "q" in pdq-notation.

genT
a boolean value to specify for a generalized t-student garch model.

asym
a string value for the asymmetric function for an asymmetric GARCH process.

chains
An integer of the number of Markov Chains chains to be run, by default 4 chains are run.

iter
An integer of total iterations per chain including the warm-up, by default the number of iterations are 2000.

warmup
A positive integer specifying number of warm-up (aka burn-in) iterations. This also specifies the number of iterations used for step-size adaptation, so warm-up samples should not be used for inference. The number of warmup should not be larger than iter and the default is iter/2.

adapt.delta
An optional real value between 0 and 1, the thin of the jumps in a HMC method. By default is 0.9

tree.depth
An integer of the maximum depth of the trees evaluated during each iteration. By default is 10.

seed
An integer with the seed for using when predicting with the model.

...
Additional arguments passed to forecast::Arima

Value

A modeltime model