Interface for Generalized Additive Models (GAM)

Interface for Generalized Additive Models (GAM)

gen_additive_reg(
  mode = "regression",
  markov_chains = NULL,
  chain_iter = NULL,
  warmup_iter = NULL,
  adapt_delta = NULL
)

Arguments

mode
A single character string for the type of model.

markov_chains
Number of Markov chains (defaults to 4).

chain_iter
Number of total iterations per chain (including warmup; defaults to 2000).

warmup_iter
A positive integer specifying number of warmup (aka burnin) iterations. This also specifies the number of iterations used for stepsize adaptation, so warmup samples should not be used for inference. The number of warmup should not be larger than iter and the default is iter/2.

adapt_delta
The thin of the jumps in a HMC method.

Value

A parsnip model specification

Details

Available Engines:

stan: Connects to brms::brm()

Engine Details

stan

This engine uses brms::brm() and has the following parameters, which can be modified through the parsnip::set_engine() function.

## function (formula, data, family = gaussian(), prior = NULL, autocor = NULL, data2 = NULL, cov_ranef = NULL, sample_prior = "no", 
##     sparse = NULL, knots = NULL, stanvars = NULL, stan_funs = NULL, fit = NA, save_pars = NULL, save_ranef = NULL, save_mevars = NULL, 
##     save_all_pars = NULL, inits = "random", chains = 4, iter = 2000, warmup = floor(iter/2), thin = 1, cores = getOption("mc.cores", 
##         1), threads = NULL, normalize = getOption("brms.normalize", TRUE), control = NULL, algorithm = getOption("brms.algorithm", 
##         "sampling"), backend = getOption("brms.backend", "rstan"), future = getOption("future", FALSE), silent = 1, seed = NA, save_model = NULL, 
##     stan_model_args = list(), file = NULL, file_refit = "never", empty = FALSE, rename = TRUE, ...)

Fit Details

BRMS Formula Interface

Fitting GAMs is accomplished using parameters including:

brms::s(): GAM spline smooths

brms::t2(): GAM tensor product smooths

These are applied in the fit() function:

fit(value ~ s(date_mon, k = 12) + s(date_num), data = df)

Examples

# \donttest{
library(tidymodels)

#> ── Attaching packages ────────────────────────────────────── tidymodels 0.1.3 ──

#> ✔ broom        0.7.6      ✔ tibble       3.1.1 
#> ✔ dials        0.0.9      ✔ tidyr        1.1.3 
#> ✔ ggplot2      3.3.3      ✔ tune         0.1.5 
#> ✔ infer        0.5.4      ✔ workflows    0.2.2 
#> ✔ modeldata    0.1.0      ✔ workflowsets 0.0.2 
#> ✔ purrr        0.3.4      ✔ yardstick    0.0.8 
#> ✔ recipes      0.1.16     

#> ── Conflicts ───────────────────────────────────────── tidymodels_conflicts() ──
#> ✖ purrr::discard()    masks scales::discard()
#> ✖ dplyr::filter()     masks stats::filter()
#> ✖ dplyr::first()      masks xts::first()
#> ✖ dplyr::lag()        masks stats::lag()
#> ✖ dplyr::last()       masks xts::last()
#> ✖ dplyr::select()     masks MASS::select()
#> ✖ recipes::step()     masks stats::step()
#> ✖ dials::tree_depth() masks bayesmodels::tree_depth()
#> • Use tidymodels_prefer() to resolve common conflicts.

library(bayesmodels)
library(modeltime)
library(tidyverse)

#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.1 ──

#> ✔ readr   1.4.0     ✔ forcats 0.5.1
#> ✔ stringr 1.4.0     

#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ lubridate::as.difftime() masks base::as.difftime()
#> ✖ readr::col_factor()      masks scales::col_factor()
#> ✖ lubridate::date()        masks base::date()
#> ✖ purrr::discard()         masks scales::discard()
#> ✖ dplyr::filter()          masks stats::filter()
#> ✖ dplyr::first()           masks xts::first()
#> ✖ stringr::fixed()         masks recipes::fixed()
#> ✖ lubridate::intersect()   masks base::intersect()
#> ✖ dplyr::lag()             masks stats::lag()
#> ✖ dplyr::last()            masks xts::last()
#> ✖ dplyr::select()          masks MASS::select()
#> ✖ lubridate::setdiff()     masks base::setdiff()
#> ✖ readr::spec()            masks yardstick::spec()
#> ✖ lubridate::union()       masks base::union()

library(timetk)
library(lubridate)

m750_extended <- m750 %>%
    group_by(id) %>%
    future_frame(.length_out = 24, .bind_data = TRUE) %>%
    mutate(lag_24 = lag(value, 24)) %>%
    ungroup() %>%
    mutate(date_num = as.numeric(date)) %>%
    mutate(date_month = month(date))

#> .date_var is missing. Using: date

m750_train  <- m750_extended %>% drop_na()
m750_future <- m750_extended %>% filter(is.na(value))

model_fit_gam <- gen_additive_reg(mode = "regression", markov_chains = 2) %>%
    set_engine("stan", family=Gamma(link="log")) %>%
    fit(value ~  date + s(date_month, k = 12)
        + s(lag_24),
        data = m750_train) 

#> frequency = 12 observations per 1 year

#> Compiling Stan program...

#> Start sampling

#> 
#> SAMPLING FOR MODEL '827d50e4850f94b3ae8c23dea39c615d' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000182 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.82 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 26.2782 seconds (Warm-up)
#> Chain 1:                26.9416 seconds (Sampling)
#> Chain 1:                53.2199 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL '827d50e4850f94b3ae8c23dea39c615d' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 9.7e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.97 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 25.5031 seconds (Warm-up)
#> Chain 2:                24.7101 seconds (Sampling)
#> Chain 2:                50.2131 seconds (Total)
#> Chain 2: 

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

 # }