General Interface for Naive and Random Walk models Regression Models

random_walk_reg() is a way to generate a specification of Naive and Random Walk models before fitting and allows the model to be created using different packages. Currently the only package is bayesforecast.

random_walk_reg(
  mode = "regression",
  seasonal_random_walk = NULL,
  seasonal_period = NULL,
  markov_chains = NULL,
  chain_iter = NULL,
  warmup_iter = NULL,
  adapt_delta = NULL,
  tree_depth = NULL,
  pred_seed = NULL
)

Arguments

mode

A single character string for the type of model. The only possible value for this model is "regression".

seasonal_random_walk

a Boolean value for select a seasonal random walk instead.

seasonal_period

an optional integer value for the seasonal period.

markov_chains

An integer of the number of Markov Chains chains to be run, by default 4 chains are run.

chain_iter

An integer of total iterations per chain including the warm-up, by default the number of iterations are 2000.

warmup_iter

A positive integer specifying number of warm-up (aka burn-in) iterations. This also specifies the number of iterations used for step-size adaptation, so warm-up samples should not be used for inference. The number of warmup should not be larger than iter and the default is iter/2.

adapt_delta

An optional real value between 0 and 1, the thin of the jumps in a HMC method. By default is 0.9

tree_depth

An integer of the maximum depth of the trees evaluated during each iteration. By default is 10.

pred_seed

An integer with the seed for using when predicting with the model.

Value

A model spec

Details

The data given to the function are not saved and are only used to determine the mode of the model. For random_walk_reg(), the mode will always be "regression".

The model can be created using the fit() function using the following engines:

"stan" (default) - Connects to bayesforecast::stan_naive()

Main Arguments

The main arguments (tuning parameters) for the model are:

seasonal_random_walk: a Boolean value for select a seasonal random walk instead.

markov_chains: An integer of the number of Markov Chains chains to be run.

adapt_delta: The thin of the jumps in a HMC method.

tree_depth: The maximum depth of the trees evaluated during each iteration.

These arguments are converted to their specific names at the time that the model is fit.

Other options and argument can be set using set_engine() (See Engine Details below).

If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

Engine Details

The standardized parameter names in bayesmodels can be mapped to their original names in each engine:

bayesmodels	bayesforecast::stan_naive

seasonal_random_walk	seasonal

markov_chains	chains(4)

adapt_delta	adapt.delta(0.9)

tree_depth	tree.depth(10)

Other options can be set using set_engine().

stam (default engine)

The engine uses bayesforecast::stan_naive().

Fit Details

Date and Date-Time Variable

It's a requirement to have a date or date-time variable as a predictor. The fit() interface accepts date and date-time features and handles them internally.

fit(y ~ date)

Seasonal Period Specification

The period can be non-seasonal (seasonal_period = 1 or "none") or yearly seasonal (e.g. For monthly time stamps, seasonal_period = 12, seasonal_period = "12 months", or seasonal_period = "yearly"). There are 3 ways to specify:

seasonal_period = "auto": A seasonal period is selected based on the periodicity of the data (e.g. 12 if monthly)

seasonal_period = 12: A numeric frequency. For example, 12 is common for monthly data

seasonal_period = "1 year": A time-based phrase. For example, "1 year" would convert to 12 for monthly data.

Univariate (No xregs, Exogenous Regressors):

For univariate analysis, you must include a date or date-time feature. Simply use:

Formula Interface (recommended): fit(y ~ date) will ignore xreg's.

See also

Examples

if (FALSE) {
library(dplyr)
library(parsnip)
library(rsample)
library(timetk)
library(modeltime)
library(bayesmodels)

# Data
m750 <- m4_monthly %>% filter(id == "M750")
m750

# Split Data 80/20
splits <- rsample::initial_time_split(m750, prop = 0.8)

# Model Spec
model_spec <- random_walk_reg() %>%
    set_engine("stan")

# Fit Spec
model_fit <- model_spec %>%
    fit(log(value) ~ date, data = training(splits))
model_fit
}