General Interface for ARIMA Regression Models

sarima_reg() is a way to generate a specification of an ARIMA model before fitting and allows the model to be created using different packages. Currently the only package is bayesforecast.

sarima_reg(
  mode = "regression",
  seasonal_period = NULL,
  non_seasonal_ar = NULL,
  non_seasonal_differences = NULL,
  non_seasonal_ma = NULL,
  seasonal_ar = NULL,
  seasonal_differences = NULL,
  seasonal_ma = NULL,
  markov_chains = NULL,
  chain_iter = NULL,
  warmup_iter = NULL,
  adapt_delta = NULL,
  tree_depth = NULL,
  pred_seed = NULL
)

Arguments

mode
A single character string for the type of model. The only possible value for this model is "regression".

seasonal_period
A seasonal frequency. Uses "auto" by default. A character phrase of "auto" or time-based phrase of "2 weeks" can be used if a date or date-time variable is provided. See Fit Details below.

non_seasonal_ar
The order of the non-seasonal auto-regressive (AR) terms. Often denoted "p" in pdq-notation.

non_seasonal_differences
The order of integration for non-seasonal differencing. Often denoted "d" in pdq-notation.

non_seasonal_ma
The order of the non-seasonal moving average (MA) terms. Often denoted "q" in pdq-notation.

seasonal_ar
The order of the seasonal auto-regressive (SAR) terms. Often denoted "P" in PDQ-notation.

seasonal_differences
The order of integration for seasonal differencing. Often denoted "D" in PDQ-notation.

seasonal_ma
The order of the seasonal moving average (SMA) terms. Often denoted "Q" in PDQ-notation.

markov_chains
An integer of the number of Markov Chains chains to be run, by default 4 chains are run.

chain_iter
An integer of total iterations per chain including the warm-up, by default the number of iterations are 2000.

warmup_iter
A positive integer specifying number of warm-up (aka burn-in) iterations. This also specifies the number of iterations used for step-size adaptation, so warm-up samples should not be used for inference. The number of warmup should not be larger than iter and the default is iter/2.

adapt_delta
An optional real value between 0 and 1, the thin of the jumps in a HMC method. By default is 0.9

tree_depth
An integer of the maximum depth of the trees evaluated during each iteration. By default is 10.

pred_seed
An integer with the seed for using when predicting with the model.

Details

The data given to the function are not saved and are only used to determine the mode of the model. For sarima_reg(), the mode will always be "regression".

The model can be created using the fit() function using the following engines:

"stan" (default) - Connects to bayesforecast::stan_sarima()

Main Arguments

The main arguments (tuning parameters) for the model are:

non_seasonal_ar: The order of the non-seasonal auto-regressive (AR) terms.

non_seasonal_differences: The order of integration for non-seasonal differencing.

non_seasonal_ma: The order of the non-seasonal moving average (MA) terms.

seasonal_ar: The order of the seasonal auto-regressive (SAR) terms.

seasonal_differences: The order of integration for seasonal differencing.

seasonal_ma: The order of the seasonal moving average (SMA) terms.

markov_chains: An integer of the number of Markov Chains chains to be run.

adapt_delta: The thin of the jumps in a HMC method.

tree_depth: The maximum depth of the trees evaluated during each iteration

These arguments are converted to their specific names at the time that the model is fit.

Other options and argument can be set using set_engine() (See Engine Details below).

If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

Engine Details

The standardized parameter names in bayesmodels can be mapped to their original names in the engine:

bayesmodels	bayesforecast::stan_sarima

non_seasonal_ar, non_seasonal_differences, non_seasonal_ma	order = c(p(1), d(0), q(0))

seasonal_ar, seasonal_differences, seasonal_ma	seasonal = c(P(0), D(0), Q(0))

markov_chains	chains(4)

adapt_delta	adapt.delta(0.9)

tree_depth	tree.depth(10)

Other options can be set using set_engine().

stan (default engine)

The engine uses bayesforecast::stan_sarima().

Parameter Notes:

xreg - This is supplied via the parsnip / bayesmodels fit() interface (so don't provide this manually). See Fit Details (below).

Fit Details

Date and Date-Time Variable

It's a requirement to have a date or date-time variable as a predictor. The fit() interface accepts date and date-time features and handles them internally.

fit(y ~ date)

Seasonal Period Specification

The period can be non-seasonal (seasonal_period = 1 or "none") or yearly seasonal (e.g. For monthly time stamps, seasonal_period = 12, seasonal_period = "12 months", or seasonal_period = "yearly"). There are 3 ways to specify:

seasonal_period = "auto": A seasonal period is selected based on the periodicity of the data (e.g. 12 if monthly)

seasonal_period = 12: A numeric frequency. For example, 12 is common for monthly data

seasonal_period = "1 year": A time-based phrase. For example, "1 year" would convert to 12 for monthly data.

Univariate (No xregs, Exogenous Regressors):

For univariate analysis, you must include a date or date-time feature. Simply use:

Formula Interface: fit(y ~ date) will ignore xreg's.

Multivariate (xregs, Exogenous Regressors)

The xreg parameter is populated using the fit() function:

Only factor, ordered factor, and numeric data will be used as xregs.

Date and Date-time variables are not used as xregs

character data should be converted to factor.

Xreg Example: Suppose you have 3 features:

y (target)

date (time stamp),

month.lbl (labeled month as a ordered factor).

The month.lbl is an exogenous regressor that can be passed to the sarima_reg() using fit():

fit(y ~ date + month.lbl) will pass month.lbl on as an exogenous regressor.

Note that date or date-time class values are excluded from xreg.

See also

Examples

# \donttest{
library(dplyr)
library(parsnip)
library(rsample)
library(timetk)
library(modeltime)
library(bayesmodels)

# Data
m750 <- m4_monthly %>% filter(id == "M750")
m750

#> # A tibble: 306 x 3
#>    id    date       value
#>    <fct> <date>     <dbl>
#>  1 M750  1990-01-01  6370
#>  2 M750  1990-02-01  6430
#>  3 M750  1990-03-01  6520
#>  4 M750  1990-04-01  6580
#>  5 M750  1990-05-01  6620
#>  6 M750  1990-06-01  6690
#>  7 M750  1990-07-01  6000
#>  8 M750  1990-08-01  5450
#>  9 M750  1990-09-01  6480
#> 10 M750  1990-10-01  6820
#> # … with 296 more rows

# Split Data 80/20
splits <- rsample::initial_time_split(m750, prop = 0.8)

# ---- ARIMA ----

# Model Spec
model_spec <- sarima_reg() %>%
    set_engine("stan")

# Fit Spec
model_fit <- model_spec %>%
    fit(log(value) ~ date, data = training(splits))

#> frequency = 12 observations per 1 year

#> 
#> SAMPLING FOR MODEL 'Sarima' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.3 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.149624 seconds (Warm-up)
#> Chain 1:                0.155347 seconds (Sampling)
#> Chain 1:                0.304971 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'Sarima' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 4.3e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.43 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.150769 seconds (Warm-up)
#> Chain 2:                0.137482 seconds (Sampling)
#> Chain 2:                0.288251 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'Sarima' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 2.8e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.28 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.142298 seconds (Warm-up)
#> Chain 3:                0.145294 seconds (Sampling)
#> Chain 3:                0.287592 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'Sarima' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 2.4e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.24 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.146179 seconds (Warm-up)
#> Chain 4:                0.15351 seconds (Sampling)
#> Chain 4:                0.299689 seconds (Total)
#> Chain 4: 

model_fit

#> parsnip model object
#> 
#> Fit time:  9.5s 
#> 
#> y ~ Sarima(0,0,0) 
#> 244 observations and 1 dimension 
#> Differences: 0 seasonal Differences: 0 
#> Current observations: 244 
#>  
#>           mean     se      5%     95%      ess   Rhat
#> mu0     9.0455 0.0002  9.0277  9.0635 4079.537 1.0001
#> sigma0  0.1685 0.0001  0.1565  0.1813 3707.772 1.0003
#> loglik 89.2138 0.0157 87.2357 90.1683 3954.735 1.0001
#> 
#>  Samples were drawn using sampling(NUTS). For each parameter, ess
#>  is the effective sample size, and Rhat is the potential
#>  scale reduction factor on split chains (at convergence, Rhat = 1). 

# Model Spec
model_spec <- sarima_reg(
        seasonal_period          = 12,
        non_seasonal_ar          = 3,
        non_seasonal_differences = 1,
        non_seasonal_ma          = 3,
        seasonal_ar              = 1,
        seasonal_differences     = 0,
        seasonal_ma              = 1
    ) %>%
    set_engine("stan")

# Fit Spec
model_fit <- model_spec %>%
    fit(log(value) ~ date, data = training(splits))

#> 
#> SAMPLING FOR MODEL 'Sarima' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000138 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.38 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 13.7426 seconds (Warm-up)
#> Chain 1:                9.41207 seconds (Sampling)
#> Chain 1:                23.1546 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'Sarima' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 7.6e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.76 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 13.1744 seconds (Warm-up)
#> Chain 2:                6.21165 seconds (Sampling)
#> Chain 2:                19.386 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'Sarima' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 8.1e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.81 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 12.6856 seconds (Warm-up)
#> Chain 3:                5.27995 seconds (Sampling)
#> Chain 3:                17.9656 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'Sarima' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 7.5e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.75 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 15.3968 seconds (Warm-up)
#> Chain 4:                4.52533 seconds (Sampling)
#> Chain 4:                19.9222 seconds (Total)
#> Chain 4: 

#> Warning: There were 82 divergent transitions after warmup. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.

#> Warning: Examine the pairs() plot to diagnose sampling problems

model_fit

#> parsnip model object
#> 
#> Fit time:  1m 28.6s 
#> 
#> y ~ Sarima(3,1,3)(1,0,1)[12] 
#> 244 observations and 1 dimension 
#> Differences: 1 seasonal Differences: 0 
#> Current observations: 243 
#>  
#>            mean     se       5%      95%      ess   Rhat
#> mu0      0.0002 0.0000  -0.0004   0.0008 3846.693 0.9998
#> sigma0   0.0232 0.0000   0.0215   0.0251 3999.412 1.0002
#> ar.1     0.0163 0.0006  -0.0435   0.0787 3380.780 1.0009
#> ar.2     0.0072 0.0006  -0.0551   0.0642 3799.892 1.0013
#> ar.3    -0.0161 0.0005  -0.0684   0.0377 3559.245 1.0011
#> ma.1    -0.2006 0.0013  -0.3375  -0.0740 3812.499 1.0015
#> ma.2    -0.2290 0.0012  -0.3443  -0.1026 3747.015 1.0028
#> ma.3    -0.1170 0.0011  -0.2277  -0.0052 4046.369 0.9999
#> sar      0.9180 0.0006   0.8534   0.9747 3889.423 1.0000
#> sma     -0.2742 0.0011  -0.3910  -0.1555 3815.635 0.9998
#> loglik 569.7324 0.0385 565.1518 573.0690 3968.396 1.0007
#> 
#>  Samples were drawn using sampling(NUTS). For each parameter, ess
#>  is the effective sample size, and Rhat is the potential
#>  scale reduction factor on split chains (at convergence, Rhat = 1). 

# }